thesis-omxspi-pipeline/requirements.txt
lseg-data>=2.0
pandas>=2.2
numpy>=1.26
pyarrow>=15.0
python-dotenv>=1.0
rapidfuzz>=3.9
tqdm>=4.66
